refractor: move functions to helpers.ts
